+++ Actions/process_symbols.cpp
@@ -7,24 +7,24 @@
 #include "../Registers/registers_symbols.h"
 #include "process_symbols.h"
 
-std::vector<unsigned long long> calling_tree;
+std::vector<size_t> calling_tree;
 
 void p_inverseJmpSign(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 	*(registers->jmp_sign) *= -1;
 }
 void p_jmp(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
-	long long count = (unsigned long long)registers->ax->get() * *(registers->jmp_sign);
+	long long count = (size_t)registers->ax->get() * *(registers->jmp_sign);
 	*(registers->process_step) += count;
 }
 void p_cmp(std::shared_ptr<void> unused_p, regs* registers, memory* mem) {
-	unsigned long long saved_rax = registers->rax->get();
-	unsigned long long saved_rbx = registers->rax->get();
+	size_t saved_rax = registers->rax->get();
+	size_t saved_rbx = registers->rax->get();
 
 	popMem(std::make_shared<registries_def>(registries_def::RAX), registers, mem);
 	popMem(std::make_shared<registries_def>(registries_def::RBX), registers, mem);
 
-	unsigned long long rax = registers->rax->get();
-	unsigned long long rbx = registers->rbx->get();
+	size_t rax = registers->rax->get();
+	size_t rbx = registers->rbx->get();
 
 	if (rax == rbx) {
 		*(registers->cmp_out) = 0;
@@ -79,7 +79,7 @@ void p_cmpstr(std::shared_ptr<void> unused_p, regs* registers, memory* mem) {
 // Jump if equal
 void p_je(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 	if (*(registers->cmp_out) == 0) {
-		long long count = (unsigned long long)registers->ax->get() * *(registers->jmp_sign);
+		long long count = (size_t)registers->ax->get() * *(registers->jmp_sign);
 		*(registers->process_step) += count;
 		*(registers->cmp_out) = 0xFF;
 	}
@@ -87,7 +87,7 @@ void p_je(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 // Jump if not equal
 void p_jne(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 	if (*(registers->cmp_out) != 0) {
-		long long count = (unsigned long long)registers->ax->get() * *(registers->jmp_sign);
+		long long count = (size_t)registers->ax->get() * *(registers->jmp_sign);
 		*(registers->process_step) += count;
 		*(registers->cmp_out) = 0xFF;
 	}
@@ -95,7 +95,7 @@ void p_jne(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 // Jump if less
 void p_jl(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 	if (*(registers->cmp_out) == 1) {
-		long long count = (unsigned long long)registers->ax->get() * *(registers->jmp_sign);
+		long long count = (size_t)registers->ax->get() * *(registers->jmp_sign);
 		*(registers->process_step) += count;
 		*(registers->cmp_out) = 0xFF;
 	}
@@ -103,7 +103,7 @@ void p_jl(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 // Jump if greater
 void p_jg(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 	if (*(registers->cmp_out) == 2) {
-		long long count = (unsigned long long)registers->ax->get() * *(registers->jmp_sign);
+		long long count = (size_t)registers->ax->get() * *(registers->jmp_sign);
 		*(registers->process_step) += count;
 		*(registers->cmp_out) = 0xFF;
 	}
@@ -111,7 +111,7 @@ void p_jg(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 // Jump if less or equal
 void p_jle(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 	if (*(registers->cmp_out) == 0 || *(registers->cmp_out) == 1) {
-		long long count = (unsigned long long)registers->ax->get() * *(registers->jmp_sign);
+		long long count = (size_t)registers->ax->get() * *(registers->jmp_sign);
 		*(registers->process_step) += count;
 		*(registers->cmp_out) = 0xFF;
 	}
@@ -119,7 +119,7 @@ void p_jle(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 //Jump if greater or equal
 void p_jge(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 	if (*(registers->cmp_out) == 0 || *(registers->cmp_out) == 2) {
-		long long count = (unsigned long long)registers->ax->get() * *(registers->jmp_sign);
+		long long count = (size_t)registers->ax->get() * *(registers->jmp_sign);
 		*(registers->process_step) += count;
 		*(registers->cmp_out) = 0xFF;
 	}
@@ -127,7 +127,7 @@ void p_jge(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
 
 // Get Current working Adress [GCA] (adress of the current action) Put it on stack memory
 void p_gca(std::shared_ptr<void> unused_p, regs* registers, memory* mem) {
-	unsigned long long saved_rax = registers->rax->get();
+	size_t saved_rax = registers->rax->get();
 
 	registers->rax->set(*(registers->process_step));
 	pushMem(std::make_shared<registries_def>(registries_def::RAX), registers, mem);
@@ -143,11 +143,11 @@ void p_hlt(std::shared_ptr<void> unused_p, regs* registers, memory* unsused_m) {
 // Get immediatly to pointed address (mainly used with tags and procedure)
 void p_call(std::shared_ptr<void> value_ptr, regs* registers, memory* unused_m) {
 	*registers->process_call_address = *registers->process_step;
-	*registers->process_step = *std::static_pointer_cast<unsigned long long>(value_ptr);
+	*registers->process_step = *std::static_pointer_cast<size_t>(value_ptr);
 }
 // Get immediatly to pointed address without saving caller address (mainly used with tags in loops)
 void p_lcall(std::shared_ptr<void> value_ptr, regs* registers, memory* unused_m) {
-	*registers->process_step = *std::static_pointer_cast<unsigned long long>(value_ptr);
+	*registers->process_step = *std::static_pointer_cast<size_t>(value_ptr);
 }
 // Get back to caller address (should only be used with p_call)
 void p_ret(std::shared_ptr<void> unused_p, regs* registers, memory* unused_m) {
