+++ Compiler/variables.h
@@ -19,7 +19,7 @@ struct code_file_decl_form {
 };
 struct tag_decl_form {
 	std::string tagname;
-	unsigned long long value = 0;
+	size_t value = 0;
 };
 
 struct variables_decl {
@@ -27,7 +27,7 @@ public:
 	variables_decl();
 	
 	void set(std::string var_name, unsigned char value[]);
-	void set(std::string var_name, unsigned char value[], unsigned long long length);
+	void set(std::string var_name, unsigned char value[], size_t length);
 
 	unsigned char* get(std::string var_name);
 	void deleteVar(std::string var_name);
@@ -37,22 +37,22 @@ public:
 
 	std::vector<code_file_decl_form> getVariablesTree();
 
-	void setTag(std::string tagname, unsigned long long value);
-	unsigned long long getTag(std::string tagname);
-	void setTagsMap(std::map<std::string, unsigned long long> map);
-	std::map<std::string, unsigned long long> getTagsMap();
+	void setTag(std::string tagname, size_t value);
+	size_t getTag(std::string tagname);
+	void setTagsMap(std::map<std::string, size_t> map);
+	std::map<std::string, size_t> getTagsMap();
 
 	void setTagsBranch(tag_decl_form branch);
 	tag_decl_form getTagBranch(std::string tagname);
 	std::vector<tag_decl_form> getTagsTree();
 
-	unsigned long long sys_vars_count = 0;
+	size_t sys_vars_count = 0;
 private:
 	std::map<std::string, m_container> vars;
 	std::map<std::string, code_file_decl_form> variables_tree;
 	std::vector<code_file_decl_form> variables_vector;
 
-	std::map<std::string, unsigned long long> tags;
+	std::map<std::string, size_t> tags;
 	std::map<std::string, tag_decl_form> tags_tree;
 	std::vector<tag_decl_form> tags_vector;
 };
